+++ TinkoffASDKCore/TinkoffASDKCore/AcquiringSdkConfiguration.swift
@@ -32,6 +32,17 @@ public enum AcquiringSdkEnvironment: String {
     case prod = "securepay.tinkoff.ru"
 }
 
+public struct ConfigSdkEnvironment: RawRepresentable {
+    public let rawValue: String
+
+    public init(rawValue: String) {
+        self.rawValue = rawValue
+    }
+    
+    public static var test: ConfigSdkEnvironment { ConfigSdkEnvironment(rawValue: "ca-3ds2-test.cdn-tinkoff.ru") }
+    public static var prod: ConfigSdkEnvironment { ConfigSdkEnvironment(rawValue: "ca-3ds2-prod.cdn-tinkoff.ru") }
+}
+
 public struct AcquiringSdkCredential {
     public var terminalKey: String
     public var publicKey: String
@@ -54,6 +65,8 @@ public class AcquiringSdkConfiguration: NSObject {
 
     public private(set) var serverEnvironment: AcquiringSdkEnvironment
     
+    public private(set) var configEnvironment: ConfigSdkEnvironment
+    
     public private(set) var requestsTimeoutInterval: TimeInterval
 
     /// Язык платёжной формы. На каком языке сервер будет присылать тексты ошибок клиенту
@@ -70,11 +83,12 @@ public class AcquiringSdkConfiguration: NSObject {
     /// Показывать ошибки после выполнения запроса
     public var showErrorAlert: Bool = true
     
-    /// Время в секундах, в течение которого хранится в памяти состояние доступности TinkoffPay
-    public var tinkoffPayStatusCacheLifeTime: TimeInterval = .defaultTinkoffPayStatusCacheLifeTime
-
     /// Объект, используя который SDK будет генерировать токен для передачи в запросах
     public var tokenGenerator: TokenGenerator
+
+    /// Время в секундах, в течение которого хранится в памяти состояние доступности TinkoffPay
+    public var tinkoffPayStatusCacheLifeTime: TimeInterval
+
     
     ///
     /// - Parameters:
@@ -86,27 +100,14 @@ public class AcquiringSdkConfiguration: NSObject {
     /// - Returns: AcquiringSdkConfiguration
     public init(credential: AcquiringSdkCredential,
                 server: AcquiringSdkEnvironment = .test,
-                requestsTimeoutInterval: TimeInterval,
-                tinkoffPayStatusCacheLifeTime: TimeInterval,
+                requestsTimeoutInterval: TimeInterval = 40,
+                tinkoffPayStatusCacheLifeTime: TimeInterval = 300,
                 tokenGenerator: TokenGenerator) {
         self.credential = credential
         self.requestsTimeoutInterval = requestsTimeoutInterval
+        self.tinkoffPayStatusCacheLifeTime = tinkoffPayStatusCacheLifeTime
         self.serverEnvironment = server
+        self.configEnvironment = server == .test ? .test : .prod
         self.tokenGenerator = tokenGenerator
     }
-    
-    public convenience init(credential: AcquiringSdkCredential,
-                            server: AcquiringSdkEnvironment = .test,
-                            tokenGenerator: TokenGenerator) {
-        self.init(credential: credential,
-                  server: server,
-                  requestsTimeoutInterval: .defaultRequestsTimeoutInterval,
-                  tinkoffPayStatusCacheLifeTime: .defaultTinkoffPayStatusCacheLifeTime,
-                  tokenGenerator: tokenGenerator)
-    }
-}
-
-private extension TimeInterval {
-    static let defaultRequestsTimeoutInterval: TimeInterval = 40
-    static let defaultTinkoffPayStatusCacheLifeTime: TimeInterval = 300
 }